requirements.txt
google-genai==0.10.0
Flask==3.0.2
Flask-Login==0.6.3
python-dotenv==1.0.1
supabase==2.3.4
posthog==3.5.0
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0
gunicorn==21.2.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.0.post1 